+++ src/nofunc-htbl/hashtbl0.ml
@@ -18,7 +18,7 @@ type (!'key, !'data) t =
 let clear t = Tbl.clear t.tbl
 let reset t = Tbl.reset t.tbl
 let copy t = { equal = t.equal; seeded_hash = t.seeded_hash; tbl = Tbl.copy t.tbl }
-let add t ~key ~data = Tbl.add ~seeded_hash:t.seeded_hash t.tbl key data
+let shadow t ~key ~data = Tbl.add ~seeded_hash:t.seeded_hash t.tbl key data
 let find t key = Tbl.find_opt ~equal:t.equal ~seeded_hash:t.seeded_hash t.tbl key
 let find_exn t key = Tbl.find ~equal:t.equal ~seeded_hash:t.seeded_hash t.tbl key
 let find_all t key = Tbl.find_all ~equal:t.equal ~seeded_hash:t.seeded_hash t.tbl key
@@ -45,8 +45,7 @@ let stats t = Tbl.stats t.tbl
 let to_seq t = Tbl.to_seq t.tbl
 let to_seq_keys t = Tbl.to_seq_keys t.tbl
 let to_seq_values t = Tbl.to_seq_values t.tbl
-let add_seq t seq = Tbl.add_seq ~seeded_hash:t.seeded_hash t.tbl seq
-
+let shadow_seq t seq = Tbl.add_seq ~seeded_hash:t.seeded_hash t.tbl seq
 let set_seq t seq = Tbl.replace_seq ~equal:t.equal ~seeded_hash:t.seeded_hash t.tbl seq
 
 let create (type key) (module Key : HashedType with type t = key) len =
@@ -112,3 +111,5 @@ let dyn_of_m__t (type key) (module T : Dynable with type t = key) data_to_dyn t 
 
 let replace = set
 let replace_seq = set_seq
+let add = shadow
+let add_seq = shadow_seq